+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [v2.0.3] - 2024-09-
 ### Fixed
 - `ARP_STRS_JOINED` should be empty in config file
+- Optimized History Trim
 
 ## [v2.0.2] - 2024-09-07
 ### Added